Implement fuzzy search and ranking
The regex based method to find key matches does not work very well for deeper
hierarchies and partial matching of the hierarchy levels. This approach uses the
SequenceMatcher from the difflib module to compute a score for an entry and rank
them accordingly.

This change also filters out all hidden file system entries to avoid traversing
and evaluating Git repositories which are commonly used to store pass storages.