Support OTPs while using gopass
Gopass behaves differently than pass in regards to OTPs. The command
used for pass-otp here is `pass otp code foo`, but if this exact command
is passed to gopass, it searches for a entry file with name `code`, then
chooses the closest approximation and continuously generates OTP codes
without exiting.

The correct command for gopass is `gopass otp -o foo`, the `-o` tells it
to generate a single OTP and exit.

Signed-off-by: Aseem Athale <athaleaseem@gmail.com>